k8s: Use a load balancer for release
Prior to this commit, the releases for the Kubernetes examples would use
the default cluster internal IP for the service it sets up. This means
the release was not accessible outside of the k8s cluster. This commit
fixes that by opting for using a LoadBalancer type for the service
instead so that the release URL is available and accessible.